move baseTopic to connector package
